idx-portfolio-alert/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yfinance==0.2.43
anthropic==0.34.2
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3